WIP: Nearly finished implementation of linear eqns
Renamed some typedefs, removed all fixed size matrix code except for a few
integer matrices used for permutation related work.

Added A_uu and B_u matrix to linearization code, all that remains is A_qud,
which is 75% done.

Need to update a few of the linearization methods to handle du/dt != 0 by
implementing methods which solve for du/dt giving q and u.  Once this du/dt is
known, it can then be used to properly compute linearized entries which depend
on du/dt:
A_uqc()
A_uqd()
These are quite messy, so it will probably be a while before these get
implemented, but for now it should be sufficient to only be able to compute
linearized dynamics about an equilibrium point.