Constraint forces now correct for steady motion.
I was incorrectly forming the velocity constraint equations by neglecting the
portion of the steer axis between the head and the foot.

Eigenvalues still are not matching.  Need to check steady turning condition by
implementing a function to convert Basu-Mandal coordinates to my coordinates,
if the steady turning conditions don't match then I likely have a bug in my
generalized inertia forces somewhere.

Need to implement tests to check linearized dynamics matrices.  Some is likely
present in the formulation of the generated code or in the manipulations in my
numeric side of things.