Eigenvalue test written.
All eigenvalues compare to within 1e-12 or better than those reported in
Meijaard et al. 2007.

Due to floating point representation and the way Table 1 in Meijaard et
al. has a fixed number of digits after the decimal point, rather than a fixed
number of total digits, 1e-12 is the best one can hope for as a bound on all
comparisons.  However, many of the eigenvalues compare exactly equal up to the
number of digits that are reported in their table.