Generate code using fast_cse branch of cdsousa/sympy.
New cse functionality is much faster but does result in more intermediate
terms. This commit also changes how the eigenvalue tests are performed to
eliminate issues with ordering eigenvalues in the complex plane. Briefly,
numerical round-off can cause complex number ordering schemes to switch the
order and make the tests fail. The new method creates a set of the correct
eigenvalues, and compares each computed eigenvalue against it, and if the
computed eigenvalue is within 1e-12 of the correct eigenvalue (in the complex
plane), it is considered to be correct and removes that eigenvalue from the
set. If the set is not empty by the end, then not all eigenvalues were matched
and this triggers a test failure.